requirements.txt
pywin32
PythonForWindows
pycryptodome
